src/requirements.txt
nibabel==5.3.2
ultralytics==8.3.61
pillow==11.1.0
matplotlib==3.10.0
scikit-learn==1.6.1 
segmentation-models-pytorch==0.4.0 
SimpleITK==2.4.1
loguru==0.7.3
pydantic==2.11.5
requests==2.32.3
python-dotenv==1.1.0
google-cloud-storage==3.1.0